requirements.txt
selenium
webdriver-manager
python-dotenv
requests
python-docx
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
undetected-chromedriver
psutil
tenacity
customtkinter
pyinstaller
